Just another test, this time exercising mecha for parsing day 2 2020

+++ aoc_2025/src/main.zig
@@ -21,7 +21,7 @@ pub fn main() !void {
         return;
     };
 
-    const day_num = try std.fmt.parseInt(u8, day_str, 10);
+    const day_num = try std.fmt.parseInt(i32, day_str, 10);
     out.printHeader(day_num);
     aoc.Runner.run(alloc, day_num);
 }
